+++ xdggs/tests/test_healpix.py
@@ -19,6 +19,7 @@ from xdggs.tests import (
     raise_if_dask_computes,
     requires_dask,
 )
+from xdggs.tests.strategies import ellipsoids
 
 
 # namespace class
@@ -42,6 +43,7 @@ class strategies:
             "order": cls.levels,
             "indexing_scheme": cls.indexing_schemes,
             "nest": st.booleans(),
+            "ellipsoid": ellipsoids("serialized_only"),
         }
 
         names = {
@@ -73,11 +75,13 @@ class strategies:
     def grids(
         levels=levels,
         indexing_schemes=indexing_schemes,
+        ellipsoids=ellipsoids("in_memory_only"),
     ):
         return st.builds(
             healpix.HealpixInfo,
             level=levels,
             indexing_scheme=indexing_schemes,
+            ellipsoid=ellipsoids,
         )
 
     @classmethod
@@ -86,6 +90,7 @@ class strategies:
         levels=levels,
         indexing_schemes=indexing_schemes,
         dtypes=None,
+        ellipsoids=ellipsoids("in_memory_only"),
     ):
         cell_levels = st.shared(levels, key="common-levels")
         grid_levels = st.shared(levels, key="common-levels")
@@ -95,6 +100,7 @@ class strategies:
         grids_ = cls.grids(
             levels=grid_levels,
             indexing_schemes=indexing_schemes,
+            ellipsoids=ellipsoids,
         )
 
         return cell_ids_, grids_
@@ -139,6 +145,16 @@ variables = [
             "indexing_scheme": "nested",
         },
     ),
+    xr.Variable(
+        "cells",
+        np.array([3]),
+        {
+            "grid_name": "healpix",
+            "level": 0,
+            "indexing_scheme": "nested",
+            "ellipsoid": "WGS84",
+        },
+    ),
 ]
 variable_combinations = list(itertools.product(variables, repeat=2))
 
@@ -159,12 +175,17 @@ class TestHealpixInfo:
                 indexing_scheme=indexing_scheme,
             )
 
-    @given(strategies.levels, strategies.indexing_schemes)
-    def test_init(self, level, indexing_scheme):
-        grid = healpix.HealpixInfo(level=level, indexing_scheme=indexing_scheme)
+    @given(
+        strategies.levels, strategies.indexing_schemes, ellipsoids("serialized_only")
+    )
+    def test_init(self, level, indexing_scheme, ellipsoid):
+        grid = healpix.HealpixInfo(
+            level=level, indexing_scheme=indexing_scheme, ellipsoid=ellipsoid
+        )
 
         assert grid.level == level
         assert grid.indexing_scheme == indexing_scheme
+        assert grid.ellipsoid == ellipsoid
 
     @given(strategies.levels)
     def test_nside(self, level):
@@ -190,23 +211,38 @@ class TestHealpixInfo:
     def test_from_dict(self, mapping) -> None:
         healpix.HealpixInfo.from_dict(mapping)
 
-    @given(strategies.levels, strategies.indexing_schemes)
-    def test_to_dict(self, level, indexing_scheme) -> None:
-        grid = healpix.HealpixInfo(level=level, indexing_scheme=indexing_scheme)
+    @given(strategies.levels, strategies.indexing_schemes, ellipsoids("in_memory_only"))
+    def test_to_dict(self, level, indexing_scheme, ellipsoid) -> None:
+        grid = healpix.HealpixInfo(
+            level=level, indexing_scheme=indexing_scheme, ellipsoid=ellipsoid
+        )
         actual = grid.to_dict()
 
-        assert set(actual) == {"grid_name", "level", "indexing_scheme"}
+        expected_names = {"grid_name", "level", "indexing_scheme"}
+        if ellipsoid is not None:
+            expected_names.add("ellipsoid")
+
+        assert set(actual) == expected_names
         assert actual["grid_name"] == "healpix"
         assert actual["level"] == level
         assert actual["indexing_scheme"] == indexing_scheme
+        if ellipsoid is not None:
+            expected_ellipsoid = (
+                ellipsoid if isinstance(ellipsoid, str) else ellipsoid.to_dict()
+            )
+            assert actual["ellipsoid"] == expected_ellipsoid
 
-    @given(strategies.levels, strategies.indexing_schemes)
-    def test_roundtrip(self, level, indexing_scheme):
+    @given(
+        strategies.levels, strategies.indexing_schemes, ellipsoids("serialized_only")
+    )
+    def test_roundtrip(self, level, indexing_scheme, ellipsoid):
         mapping = {
             "grid_name": "healpix",
             "level": level,
             "indexing_scheme": indexing_scheme,
         }
+        if ellipsoid is not None:
+            mapping["ellipsoid"] = ellipsoid
 
         grid = healpix.HealpixInfo.from_dict(mapping)
         roundtripped = grid.to_dict()
@@ -280,6 +316,18 @@ class TestHealpixInfo:
                     ]
                 ),
             ),
+            (
+                {"level": 2, "indexing_scheme": "nested", "ellipsoid": "WGS84"},
+                np.array([79]),
+                np.array(
+                    [
+                        [0.0, 19.55202227],
+                        [11.25, 30.11125172],
+                        [0.0, 41.93785391],
+                        [-11.25, 30.11125172],
+                    ]
+                ),
+            ),
         ),
     )
     @pytest.mark.parametrize("backend", ["shapely", "geoarrow"])
@@ -314,29 +362,53 @@ class TestHealpixInfo:
         np.testing.assert_equal(roundtripped, cell_ids)
 
     @pytest.mark.parametrize(
-        ["cell_ids", "level", "indexing_scheme", "expected"],
+        ["cell_ids", "level", "indexing_scheme", "ellipsoid", "expected"],
         (
             pytest.param(
                 np.array([3]),
                 1,
                 "ring",
+                None,
                 (np.array([315.0]), np.array([66.44353569089877])),
             ),
             pytest.param(
                 np.array([5, 11, 21]),
                 3,
                 "nested",
+                None,
                 (
                     np.array([61.875, 33.75, 84.375]),
                     np.array([19.47122063, 24.62431835, 41.8103149]),
                 ),
             ),
+            pytest.param(
+                np.array([3]),
+                1,
+                "ring",
+                "bessel",
+                (
+                    np.array([315.0]),
+                    np.array([66.53709311]),
+                ),
+            ),
+            pytest.param(
+                np.array([5, 11, 21]),
+                3,
+                "nested",
+                "WGS84",
+                (
+                    np.array([61.875, 33.75, 84.375]),
+                    np.array([19.55202227, 24.72167338, 41.93785391]),
+                ),
+            ),
         ),
     )
     def test_cell_ids2geographic(
-        self, cell_ids, level, indexing_scheme, expected
+        self, cell_ids, level, indexing_scheme, ellipsoid, expected
     ) -> None:
-        grid = healpix.HealpixInfo(level=level, indexing_scheme=indexing_scheme)
+        grid = healpix.HealpixInfo(
+            level=level, indexing_scheme=indexing_scheme, ellipsoid=ellipsoid
+        )
 
         actual_lon, actual_lat = grid.cell_ids2geographic(cell_ids)
 
@@ -344,12 +416,13 @@ class TestHealpixInfo:
         np.testing.assert_allclose(actual_lat, expected[1])
 
     @pytest.mark.parametrize(
-        ["cell_centers", "level", "indexing_scheme", "expected"],
+        ["cell_centers", "level", "indexing_scheme", "ellipsoid", "expected"],
         (
             pytest.param(
                 np.array([[315.0, 66.44353569089877]]),
                 1,
                 "ring",
+                None,
                 np.array([3]),
             ),
             pytest.param(
@@ -358,14 +431,33 @@ class TestHealpixInfo:
                 ),
                 3,
                 "nested",
+                None,
+                np.array([5, 11, 21]),
+            ),
+            pytest.param(
+                np.array([[315.0, 66.53709311]]),
+                1,
+                "ring",
+                "bessel",
+                np.array([3]),
+            ),
+            pytest.param(
+                np.array(
+                    [[61.875, 19.55202227], [33.75, 24.72167338], [84.375, 41.93785391]]
+                ),
+                3,
+                "nested",
+                "WGS84",
                 np.array([5, 11, 21]),
             ),
         ),
     )
     def test_geographic2cell_ids(
-        self, cell_centers, level, indexing_scheme, expected
+        self, cell_centers, level, indexing_scheme, ellipsoid, expected
     ) -> None:
-        grid = healpix.HealpixInfo(level=level, indexing_scheme=indexing_scheme)
+        grid = healpix.HealpixInfo(
+            level=level, indexing_scheme=indexing_scheme, ellipsoid=ellipsoid
+        )
 
         actual = grid.geographic2cell_ids(
             lon=cell_centers[:, 0], lat=cell_centers[:, 1]