+++ docs/changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.4.1 (_unreleased_)
+## 0.5.0 (2025-12-12)
+
+- Support ellipsoids for `healpix` ({pull}`194`)
+- Move the plotting code into a custom `anywidget`-based widget ({pull}`199`)
+- Temporarily remove the interactive facet plot feature ({pull}`200`)
 
 ## 0.4.0 (2025-11-03)